create sub components
